@@ -5,7 +5,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 2.0.0+5
 
 environment:
-  sdk: '>=3.0.0'
+  sdk: '^3.7.0'
 
 dependencies:
   flutter:
