@@ -7,10 +7,7 @@ class Location {
   final double longitude;
   final double latitude;
 
-  Location({
-    required this.longitude,
-    required this.latitude,
-  });
+  Location({required this.longitude, required this.latitude});
 
   factory Location.fromJson(Map<String, dynamic> json) =>
       _$LocationFromJson(json);
