@@ -7,11 +7,11 @@ part of 'location.dart';
 // **************************************************************************
 
 Location _$LocationFromJson(Map<String, dynamic> json) => Location(
-      longitude: (json['longitude'] as num).toDouble(),
-      latitude: (json['latitude'] as num).toDouble(),
-    );
+  longitude: (json['longitude'] as num).toDouble(),
+  latitude: (json['latitude'] as num).toDouble(),
+);
 
 Map<String, dynamic> _$LocationToJson(Location instance) => <String, dynamic>{
-      'longitude': instance.longitude,
-      'latitude': instance.latitude,
-    };
+  'longitude': instance.longitude,
+  'latitude': instance.latitude,
+};
